+++ src/main/java/kr/mywork/common/auth/config/JwtProperties.java
@@ -1,4 +1,4 @@
-package kr.mywork.common.auth;
+package kr.mywork.common.auth.config;
 
 import lombok.Getter;
 import lombok.Setter;
